+++ src/enums/event-types.ts
@@ -1,0 +1,10 @@
+/**
+ * Event types
+ * 
+ * @enum {string}
+ */
+export const enum EventType {
+    START = 'Start',
+    STOP = 'Stop',
+    COMMAND = 'Command'
+}