+++ CHANGELOG.md
@@ -1,5 +1,59 @@
 # @reveldigital/client-sdk
 
+## 1.0.0-alpha.6
+
+### Patch Changes
+
+- export options interface
+
+## 1.0.0-alpha.5
+
+### Patch Changes
+
+- add legacy event support
+
+## 1.0.0-alpha.4
+
+### Patch Changes
+
+- add gadget types dependency
+
+## 1.0.0-alpha.3
+
+### Patch Changes
+
+- refactor event listeners
+
+## 1.0.0-alpha.2
+
+### Patch Changes
+
+- bump
+
+## 1.0.1-alpha.1
+
+### Patch Changes
+
+- none
+
+## 1.0.1-alpha.0
+
+### Patch Changes
+
+- alpha release
+
+## 1.0.0
+
+### Major Changes
+
+- initial changeset
+
+## 0.0.1
+
+### Patch Changes
+
+- first release
+
 ## 1.0.0
 
 ### Patch Changes